# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Interview Guide Platform API",
    "description": "High-performance Rust Axum API for Interview Guide & Technical Knowledge Platform",
    "contact": {
      "name": "Platform Engineering",
      "email": "eng@interviewguide.internal"
    },
    "license": {
      "name": ""
    },
    "version": "2.0.0"
  },
  "paths": {
    "/api/audit/events": {
      "get": {
        "tags": [
          "Enterprise"
        ],
        "summary": "GET /api/audit/events",
        "operationId": "list_events",
        "responses": {
          "200": {
            "description": "List recent platform audit ledger events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEventItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Admin access required"
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign in with email and password.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed in; refresh cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Account disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Too many failed attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Current signed-in user's profile.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Exchange the HTTP-only refresh cookie for a new access token (rotates the refresh token).",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "New access token; refresh cookie rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create an account and start a session.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created; refresh cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEventItem": {
        "type": "object",
        "required": [
          "sequence_id",
          "event_type",
          "payload_json",
          "prev_hash",
          "curr_hash",
          "created_at"
        ],
        "properties": {
          "actor_id": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "curr_hash": {
            "type": "string"
          },
          "event_type": {
            "type": "string"
          },
          "payload_json": {},
          "prev_hash": {
            "type": "string"
          },
          "sequence_id": {
            "type": "integer",
            "format": "int64"
          },
          "target_id": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "user"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "description": "Access token lifetime in seconds. Refresh via `POST /api/auth/refresh` (HTTP-only cookie)."
          },
          "token_type": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfile"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "full_name"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "full_name": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}